requirements.txt
scikit-learn
psycopg2-binary
pandas
dtaidistance
sqlalchemy
faker